requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
pydantic==2.10.6
google-genai==1.3.0
numpy==2.0.2
Flask==3.1.0
reportlab==4.3.1